requirements.txt
numpy
Pillow
fastapi
uvicorn
python-multipart